Made traverse_and_document_shape a base class
