+++ botocore/docs/utils.py
@@ -124,59 +124,3 @@ class AutoPopulatedParam(object):
             section = section.get_section('structure-value')
             if self.name in section.available_sections:
                 section.delete_section(self.name)
-
-
-def traverse_and_document_shape(documenter, section, shape, history,
-                                include=None, exclude=None, name=None,
-                                is_required=False):
-    """Traverses and documents a shape
-
-    Will take a documenter class and call its appropriate methods as a shape
-    is traversed.
-
-    :param documenter: The documenter class to dispatch to as the shape is
-        traversed.
-
-    :param section: The section to document.
-
-    :param history: A list of the names of the shapes that have been traversed.
-
-    :type include: Dictionary where keys are parameter names and
-        values are the shapes of the parameter names.
-    :param include: The parameter shapes to include in the documentation.
-
-    :type exclude: List of the names of the parameters to exclude.
-    :param exclude: The names of the parameters to exclude from
-        documentation.
-
-    :param name: The name of the shape.
-
-    :param is_required: If the shape is a required member.
-    """
-    param_type = shape.type_name
-    if shape.name in history:
-        documenter.document_recursive_shape(section, shape, name=name)
-    else:
-        history.append(shape.name)
-        is_top_level_param = (len(history) == 2)
-        getattr(documenter, 'document_shape_type_%s' % param_type,
-                documenter.document_shape_default)(
-                    section, shape, history=history, name=name,
-                    include=include, exclude=exclude,
-                    is_top_level_param=is_top_level_param,
-                    is_required=is_required)
-        if is_top_level_param:
-            documenter.event_emitter.emit(
-                'docs.%s.%s.%s.%s' % (documenter.EVENT_NAME,
-                                      documenter.service,
-                                      documenter.operation,
-                                      name),
-                section=section)
-        at_overlying_method_section = (len(history) == 1)
-        if at_overlying_method_section:
-            documenter.event_emitter.emit(
-                'docs.%s.%s.%s.complete-section' % (documenter.EVENT_NAME,
-                                                    documenter.service,
-                                                    documenter.operation),
-                section=section)
-        history.pop()
