+++ tests/unit/docs/test_utils.py
@@ -17,12 +17,7 @@ from tests.unit.docs import BaseDocsTest
 from botocore.docs.utils import py_type_name
 from botocore.docs.utils import py_default
 from botocore.docs.utils import get_official_service_name
-from botocore.docs.utils import traverse_and_document_shape
 from botocore.docs.utils import AutoPopulatedParam
-from botocore.docs.example import ResponseExampleDocumenter
-from botocore.docs.example import RequestExampleDocumenter
-from botocore.docs.params import RequestParamsDocumenter
-from botocore.docs.params import ResponseParamsDocumenter
 
 
 class TestPythonTypeName(unittest.TestCase):
@@ -179,78 +174,3 @@ class TestAutopopulatedParam(BaseDocsTest):
         self.param.document_auto_populated_param(
             'docs.request-example', self.doc_structure)
         self.assert_contains_line(example)
-
-
-class TestTraverseAndDocumentShape(BaseDocsTest):
-    def setUp(self):
-        super(TestTraverseAndDocumentShape, self).setUp()
-        self.event_emitter = mock.Mock()
-        self.service = 'myservice'
-        self.operation = 'SampleOperation'
-        self.response_example_doc = ResponseExampleDocumenter(
-            service=self.service, operation=self.operation,
-            event_emitter=self.event_emitter)
-        self.request_example_doc = RequestExampleDocumenter(
-            service=self.service, operation=self.operation,
-            event_emitter=self.event_emitter)
-        self.response_params_doc = ResponseParamsDocumenter(
-            service=self.service, operation=self.operation,
-            event_emitter=self.event_emitter)
-        self.request_params_doc = RequestParamsDocumenter(
-            service=self.service, operation=self.operation,
-            event_emitter=self.event_emitter)
-        self.add_shape_to_params('Foo', 'String', 'This describes foo.')
-
-    def test_events_emitted_response_example(self):
-        traverse_and_document_shape(
-            documenter=self.response_example_doc, section=self.doc_structure,
-            shape=self.operation_model.input_shape, history=[]
-        )
-        structure_section = self.doc_structure.get_section('structure-value')
-        self.assertEqual(
-            self.event_emitter.emit.call_args_list,
-            [mock.call('docs.response-example.myservice.SampleOperation.Foo',
-                       section=structure_section.get_section('Foo')),
-             mock.call(('docs.response-example.myservice.SampleOperation'
-                        '.complete-section'), section=self.doc_structure)]
-        )
-
-    def test_events_emitted_request_example(self):
-        traverse_and_document_shape(
-            documenter=self.request_example_doc, section=self.doc_structure,
-            shape=self.operation_model.input_shape, history=[]
-        )
-        structure_section = self.doc_structure.get_section('structure-value')
-        self.assertEqual(
-            self.event_emitter.emit.call_args_list,
-            [mock.call('docs.request-example.myservice.SampleOperation.Foo',
-                       section=structure_section.get_section('Foo')),
-             mock.call(('docs.request-example.myservice.SampleOperation'
-                        '.complete-section'), section=self.doc_structure)]
-        )
-
-    def test_events_emitted_response_params(self):
-        traverse_and_document_shape(
-            documenter=self.response_params_doc, section=self.doc_structure,
-            shape=self.operation_model.input_shape, history=[]
-        )
-        self.assertEqual(
-            self.event_emitter.emit.call_args_list,
-            [mock.call('docs.response-params.myservice.SampleOperation.Foo',
-                       section=self.doc_structure.get_section('Foo')),
-             mock.call(('docs.response-params.myservice.SampleOperation'
-                        '.complete-section'), section=self.doc_structure)]
-        )
-
-    def test_events_emitted_request_params(self):
-        traverse_and_document_shape(
-            documenter=self.request_params_doc, section=self.doc_structure,
-            shape=self.operation_model.input_shape, history=[]
-        )
-        self.assertEqual(
-            self.event_emitter.emit.call_args_list,
-            [mock.call('docs.request-params.myservice.SampleOperation.Foo',
-                       section=self.doc_structure.get_section('Foo')),
-             mock.call(('docs.request-params.myservice.SampleOperation'
-                        '.complete-section'), section=self.doc_structure)]
-        )
